Don't overwrite fields if only carriage return is entered
In other words, prevent accidentally deleting fields. This has always
bugged me about PWman. If I edit an entry and accidentally press a
number to edit one of the fields there is no way to escape out of this.
You have to re-type or paste in the old value.

This small change means that if you just enter return then the old value
is kept.

This is good enough for me. There is one downside: if you really want to
make a field blank the only way to do it is by entering a space and then
return - but I can't think of anytime I've ever wanted to do that.