Work-around for segfault & data corruption when secring is not available
A couple of things here:

1) I probably use PWMan in a way that wasn't intended: I only make my
secret key available to in temporarily, once authorised I delete the
secring.gpg file entirely so it isn't hanging about on a machine. So
when it times out I need to remember to copy my key back to the machine
before I try to re-authorise. Sometimes I forget and ...

2) It seems that for reasons I don't have the inclination or ability to
figure out, PWMan segfaults on exit when the secring.gpg isn't available
or is 0 bytes (or, probably, doesn't contain the required secret key).
This on it's own isn't a bad thing, but when it does this it also seems
to have a nasty habit of corrupting the password database; not always in
an immediately obvious manner so I can end up "backing up" this
corrupted version.

So as a workaround (for my usage circumstances), this fix checks whether
the secring file is 0 bytes and if it is waits until it isn't (I.e. I've
copied it back to the machine) before asking for re-authorisation.

This isn't perfect as it only waits once. I.e. if it recognises the
secring is available it'll prompt for the password, but it is possible
that in this time I could get distracted and come back after the secring
file is removed again (since I remove it automatically) and forget and
cause a segfault, but then I've only my stupid self to blame.