Avoid getting stuck in a loop when moving items
I have a habit of hitting m instead of M and then getting stuck in a
loop of "Sublist does not exist, try again" because there is no sublist
I can move to. Although the code is meant to allow you to break out of
this by pressing return (the answer[0] bit) it doesn't work for me
(perhaps because of the changes I made on ui_statusline_ask_str that
restore values on return?).

So instead of running it in a loop, just run it once on keypress - it's
not exactly a hassle to run it again if I do actually want to move an
item and type it wrong.