Add ruler below password to aid visual selection of individual chars
Been meaning to do this for AGES and I just never got around to it (I
always assume doing anything in C is going to be hard and it's just not
always the case). Anyway, did today.

This just adds a ruler/counter below the password field like so:

    Password:	djdhhfhfkjhfuwhuifbvcnotrealmdsnwehbweha
                1234567890123456789012345678901234567890

Which is a huge help for logins that say something along the lines of
"Enter character 15, 23 and 31".

God knows why I still use this password manager. Securty through
obscurity I guess.